requirements.txt
streamlit
pandas
numpy
groq
python-dotenv